Add files via upload
A DataTable in UiPath is a fundamental data structure used for organizing and manipulating tabular data. It acts as an in-memory representation of a table, similar to a database table or an Excel spreadsheet. DataTables are versatile and widely employed in UiPath projects for tasks such as data extraction, manipulation, and storage.